Distribution functions and random number generation.

Probability density, cumulative distribution function, quantile function and random variate generation for the two types of Mittag-Leffler distribution. The Laplace inversion algorithm by Garrappa is used for the pdf and cdf (see https://www.mathworks.com/matlabcentral/fileexchange/48154-the-mittag-leffler-function).

dml(x, tail, scale = 1, log = FALSE, second.type = FALSE)

pml(q, tail, scale = 1, second.type = FALSE, lower.tail = TRUE, log.p = FALSE)

qml(p, tail, scale = 1, second.type = FALSE, lower.tail = TRUE, log.p = FALSE)

rml(n, tail, scale = 1, second.type = FALSE)

Arguments

x, q
vector of quantiles.

tail
tail parameter.

scale
scale parameter.

log, log.p
logical; if TRUE, probabilities p are given as log(p).

second.type
logical; if FALSE (default), first type of Mittag-Leffler distribution is assumed.

lower.tail
logical; if TRUE, probabilities are \(P[X \le x]\) otherwise, \(P[X > x]\)

p
vector of probabilities.

n
number of random draws.

Value

dml returns the density, pml returns the distribution function, qml returns the quantile function, and rml generates random variables.

Details

The Mittag-Leffler function mlf defines two types of probability distributions:

The first type of Mittag-Leffler distribution assumes the Mittag-Leffler function as its tail function, so that the CDF is given by $$F(q; \alpha, \tau) = 1 - E_{\alpha,1} (-(q/\tau)^\alpha)$$ for \(q \ge 0\), tail parameter \(0 < \alpha \le 1\), and scale parameter \(\tau > 0\). Its PDF is given by $$f(x; \alpha, \tau) = x^{\alpha - 1} E_{\alpha,\alpha} [-(x/\tau)^\alpha] / \tau^\alpha.$$ As \(\alpha\) approaches 1 from below, the Mittag-Leffler converges (weakly) to the exponential distribution. For \(0 < \alpha < 1\), it is (very) heavy-tailed, i.e. has infinite mean.

The second type of Mittag-Leffler distribution is defined via the Laplace transform of its density f: $$\int_0^\infty \exp(-sx) f(x; \alpha, 1) dx = E_{\alpha,1}(-s)$$ It is light-tailed, i.e. all its moments are finite. At scale \(\tau\), its density is $$f(x; \alpha, \tau) = f(x/\tau; \alpha, 1) / \tau.$$

References

Haubold, H. J., Mathai, A. M., & Saxena, R. K. (2011). Mittag-Leffler Functions and Their Applications. Journal of Applied Mathematics, 2011, 1–51. http://doi.org/10.1155/2011/298628

Mittag-Leffler distribution. (2017, May 3). In Wikipedia, The Free Encyclopedia. https://en.wikipedia.org/w/index.php?title=Mittag-Leffler_distribution&oldid=778429885

Examples

dml(1, 0.8)

#> [1] 0.2557438

dml(1, 0.6, second.type=TRUE)

#> [1] 0.4832354

pml(2, 0.7, 1.5)

#> [1] 0.6597308

qml(p = c(0.25, 0.5, 0.75), tail = 0.6, scale = 100)

#> [1]  11.64327  60.56854 284.67928

rml(10, 0.7, 1)

#>  [1] 0.795198647 0.408205091 0.161459719 1.067977006 0.003716904 0.126989515
#>  [7] 0.163714569 1.936521213 0.363758400 0.381153862