Log-Moments Estimator for the Mittag-Leffler Distribution (Type 1).

Tail and scale parameter of the Mittag-Leffler distribution are estimated by matching with the first two empirical log-moments (see Cahoy et al., http://doi.org/10.1016/j.jspi.2010.04.016).

logMomentEstimator(x, alpha = 0.05)

Arguments

x
A vector of non-negative data.

alpha
Confidence intervals are calculated at level 1 - alpha.

Value

A named vector with entries (nu, delta, nuLo, nuHi, deltaLo, deltaHi) where nu is the tail parameter and delta the scale parameter of the Mittag-Leffler distribution, with confidence intervals (nuLo, nuHi) resp. (deltaLo, deltaHi).

References

Cahoy, D. O., Uchaikin, V. V., & Woyczynski Wojbor, W. A. (2010). Parameter estimation for fractional Poisson processes. Journal of Statistical Planning and Inference, 140(11), 3106–3120. http://doi.org/10.1016/j.jspi.2010.04.016

Cahoy, D. O. (2013). Estimation of Mittag-Leffler Parameters. Communications in Statistics - Simulation and Computation, 42(2), 303–315. http://doi.org/10.1080/03610918.2011.640094

Examples

logMomentEstimator(rml(n = 1000, scale = 0.03, tail = 0.84), alpha=0.95)

#>       tail      scale     tailLo     tailHi    scaleLo    scaleHi 
#> 0.85682030 0.03128040 0.85571990 0.85792069 0.03117594 0.03138487