Distribution functions and random number generation.

Probability density, cumulative distribution function, quantile function and random variate generation for the two types of Mittag-Leffler distribution.

dml(x, tail, scale = 1, log = FALSE, second.type = FALSE)

pml(q, tail, scale = 1, second.type = FALSE, lower.tail = TRUE,
  log.p = FALSE)

qml(p, tail, scale = 1, second.type = FALSE, lower.tail = TRUE,
  log.p = FALSE)

rml(n, tail, scale = 1, second.type = FALSE)

Arguments

x, q
vector of quantiles.

tail
tail parameter.

scale
scale parameter.

log, log.p
logical; if TRUE, probabilities p are given as log(p).

second.type
logical; if FALSE (default), first type of Mittag-Leffler distribution is assumed.

lower.tail
logical; if TRUE, probabilities are \(P[X \le x]\) otherwise, \(P[X > x]\)

p
vector of probabilities.

n
number of observations. If length(n) > 1, the length is taken to be the number required.

Value

dml returns the density, pml returns the distribution function, qml returns the quantile function, and rml generates random variables.

Details

The Mittag-Leffler function mlf defines two types of probability distributions:

The first type of Mittag-Leffler distribution assumes the Mittag-Leffler function as its tail function, so that the CDF is given by $$F(q; \alpha, \tau) = 1 - E_{\alpha,1} (-(q/\tau)^\alpha)$$ for \(q \ge 0\), tail parameter \(0 < \alpha \le 1\), and scale parameter \(\tau > 0\). Its PDF is given by $$f(x; \alpha, \tau) = x^{\alpha - 1} E_{\alpha,\alpha} [-(x/\tau)^\alpha] / \tau^\alpha.$$ As \(\alpha\) approaches 1 from below, the Mittag-Leffler converges (weakly) to the exponential distribution. For \(0 < \alpha < 1\), it is (very) heavy-tailed, i.e. has infinite mean.

The second type of Mittag-Leffler distribution is defined via the Laplace transform of its density f: $$\int_0^\infty \exp(-sx) f(x; \alpha, 1) dx = E_{\alpha,1}(-s)$$ It is light-tailed, i.e. all its moments are finite. At scale \(\tau\), its density is $$f(x; \alpha, \tau) = f(x/\tau; \alpha, 1) / \tau.$$

References

Haubold, H. J., Mathai, A. M., & Saxena, R. K. (2011). Mittag-Leffler Functions and Their Applications. Journal of Applied Mathematics, 2011, 1–51. http://doi.org/10.1155/2011/298628

Mittag-Leffler distribution. (2017, May 3). In Wikipedia, The Free Encyclopedia. https://en.wikipedia.org/w/index.php?title=Mittag-Leffler_distribution&oldid=778429885

Examples

dml(1, 0.8)

#> [1] 0.2557438

dml(1, 0.6, second.type=TRUE)

#> [1] 0.4832354

pml(2, 0.7, 1.5)

#> [1] 0.6597308

qml(p = c(0.25, 0.5, 0.75), tail = 0.6, scale = 100)

#> [1]  11.64327  60.56854 284.67928

rml(1000, 0.7, 1)

#>    [1] 6.626348e-01 5.228339e-01 8.570974e-02 9.456973e-01 1.175239e-01
#>    [6] 4.114956e-01 6.574724e-01 6.838514e-01 7.839655e-01 2.426419e-02
#>   [11] 2.299840e-01 2.366995e-01 4.724625e-03 1.633326e+00 7.730964e-01
#>   [16] 3.509915e-01 2.199054e-01 7.123195e-01 6.870450e+00 2.192454e+00
#>   [21] 9.746584e-01 9.311020e-02 1.879563e-01 5.831034e-02 5.558061e-02
#>   [26] 5.763928e-01 1.136067e-01 9.020706e-01 4.390036e+02 2.591128e-01
#>   [31] 4.106564e-03 1.122279e+00 2.995879e-01 7.632530e-01 1.027820e-02
#>   [36] 5.108468e-01 3.169594e+00 5.424581e+00 3.215138e-02 1.543915e-01
#>   [41] 6.216023e-01 2.500721e-02 5.266553e-01 7.820486e-01 1.156833e+01
#>   [46] 2.696316e-01 1.840943e-01 1.969478e+00 6.574660e-02 9.064430e-02
#>   [51] 8.355565e-01 5.550239e+00 1.005652e-02 2.230156e+00 6.692407e-02
#>   [56] 4.078373e-03 8.299627e-02 2.890491e-01 1.127189e+00 1.828136e+00
#>   [61] 6.129399e-02 1.354323e-01 4.295576e+00 1.240069e+00 1.861366e-01
#>   [66] 2.386578e-01 1.336574e+00 7.739082e-01 8.031130e-04 3.720727e-03
#>   [71] 6.825727e+00 1.501438e+00 2.063484e+00 3.642872e-01 1.105331e+00
#>   [76] 7.723448e-02 1.422287e+00 9.726774e-02 4.154184e+00 1.564391e+01
#>   [81] 4.573007e+00 5.270772e-01 1.242937e+00 1.707752e-02 8.197444e-03
#>   [86] 1.178033e+00 2.153730e-01 5.124039e-04 5.090704e+00 6.156304e-02
#>   [91] 5.480796e-01 8.653613e-01 2.241039e-01 6.044926e-01 1.490185e-01
#>   [96] 3.627303e-01 2.867453e-01 2.943580e+00 1.704794e-01 2.210164e+00
#>  [101] 1.341747e+02 4.934873e-01 1.260226e+00 1.446904e+00 1.077825e+00
#>  [106] 8.270555e-01 1.939700e+00 5.686005e-01 1.909335e-01 7.108680e+02
#>  [111] 2.272458e-01 3.245699e+00 1.865606e+00 2.990075e+00 6.385857e-02
#>  [116] 9.298250e+00 1.900143e-01 2.798749e+00 2.191255e+00 2.334366e-01
#>  [121] 3.828030e+00 2.052328e+00 8.525319e-01 6.176748e+01 5.022583e-01
#>  [126] 3.435295e-01 1.891488e+00 5.800004e-02 3.578245e+00 6.573955e+00
#>  [131] 1.161687e-02 1.606343e-02 1.043794e+01 9.750972e-01 1.613649e-01
#>  [136] 3.755508e+00 3.478124e-02 1.402798e+00 2.725278e+00 2.310290e+00
#>  [141] 2.888857e-01 2.226776e-01 2.630209e+00 4.897685e+00 2.241056e-02
#>  [146] 2.038919e-01 2.496938e-01 3.128328e-01 4.193470e-01 2.010008e-01
#>  [151] 7.014257e+00 8.423731e-01 5.871894e-02 9.541993e-02 1.449635e+00
#>  [156] 5.845897e+00 3.183693e+00 4.940610e-01 4.877410e+01 4.832367e-02
#>  [161] 9.566816e-01 2.498559e-03 2.402634e+00 9.153378e-02 8.823012e-01
#>  [166] 1.287200e-02 2.673879e+02 1.890030e-01 7.579163e-02 3.396538e-01
#>  [171] 1.118858e+01 1.692310e-01 8.310039e-01 1.015388e+00 2.123433e+00
#>  [176] 4.694971e-01 7.249146e-01 2.696759e-01 2.431193e+00 9.289989e-01
#>  [181] 2.860810e-04 1.629802e+00 2.522132e+00 5.038577e-03 1.678239e-02
#>  [186] 6.935699e-01 1.551697e-01 3.056182e-02 8.022488e-01 5.188309e+00
#>  [191] 7.881266e+01 1.255805e+00 9.906467e-02 1.079730e+01 9.277622e-01
#>  [196] 2.327328e-02 4.559984e-01 3.014786e-01 1.135697e+00 3.046836e+00
#>  [201] 4.229617e+04 1.957798e+01 2.093814e+01 8.477518e-02 2.121359e-01
#>  [206] 1.557576e+00 4.658219e-02 3.068652e-02 3.064531e-01 6.058964e-02
#>  [211] 2.413840e-01 6.425963e-03 4.079364e+00 5.479216e+01 1.433570e-02
#>  [216] 6.251103e-01 8.109237e-01 5.601386e-02 6.069534e-01 6.963458e-02
#>  [221] 3.291541e-01 8.360754e-03 3.510245e-03 4.717266e-03 2.206584e+00
#>  [226] 1.850916e+00 8.148599e-02 1.510337e+00 8.975043e-01 1.128217e+01
#>  [231] 1.498377e+01 3.530548e-01 8.847929e-01 1.949843e-01 9.046375e-01
#>  [236] 1.452797e+00 8.382315e-02 1.021260e+00 1.409196e-01 1.163154e-01
#>  [241] 2.102694e-01 9.322553e+00 1.220660e+00 7.990773e+01 2.473675e-01
#>  [246] 5.269765e-01 8.881876e-01 9.760039e-01 1.023610e-02 1.492800e-01
#>  [251] 2.779183e+00 1.108234e+00 4.038352e-02 8.119124e-01 3.566883e-01
#>  [256] 6.255349e-01 1.709245e-01 3.845034e+00 1.433518e+00 2.925960e+00
#>  [261] 6.434531e+00 1.716300e+00 2.387631e-01 9.478034e-01 3.350489e-01
#>  [266] 1.708518e-01 4.974029e-01 5.178203e-01 2.708928e+00 1.024164e+00
#>  [271] 6.844076e+01 2.382019e-01 1.343434e+00 7.120573e-01 2.729803e-01
#>  [276] 1.314379e+00 4.566522e-02 2.456527e+00 3.517318e-01 8.931598e-03
#>  [281] 3.619064e+00 5.977323e-02 1.120630e+00 9.784294e-01 3.486682e-01
#>  [286] 4.492525e-01 4.538591e+00 5.480403e-01 2.511044e+00 3.348816e-01
#>  [291] 3.657922e-01 7.209831e+01 5.644113e-01 8.652827e+00 1.471337e-02
#>  [296] 9.848615e-01 7.595277e-03 2.353099e-02 7.168849e-01 4.747596e+00
#>  [301] 7.166420e-01 1.821212e+00 7.533561e-01 1.333754e+01 2.639243e-01
#>  [306] 4.711303e-01 3.907228e-01 1.823011e-02 8.392118e-01 5.162007e+00
#>  [311] 1.238011e-01 2.199492e-01 5.919910e+00 1.562926e-02 1.456574e+00
#>  [316] 2.782449e-03 1.207017e+00 1.171755e+00 1.021435e-01 6.413435e-02
#>  [321] 3.841494e-01 1.109904e+01 9.220627e-01 1.480291e-01 2.340561e+00
#>  [326] 1.328194e+00 2.913771e-01 2.615530e+00 2.007388e-01 5.857763e-01
#>  [331] 3.282657e-01 3.017263e+00 2.748927e+00 1.479829e-01 1.134547e-01
#>  [336] 1.246572e-01 4.418502e-01 5.915441e-02 3.146442e+00 3.707627e-01
#>  [341] 1.296190e+00 3.417885e-02 2.375385e-01 5.658188e-01 2.308087e-01
#>  [346] 6.103945e-04 6.691044e+00 1.760581e-01 4.113101e-01 4.382349e-02
#>  [351] 2.654407e-01 2.022192e-02 3.527904e-02 9.642106e-02 1.641199e-03
#>  [356] 6.333366e-01 1.215364e+00 5.054953e-01 6.511363e-01 7.335110e-01
#>  [361] 2.496479e-02 2.839155e+00 1.261996e-01 2.352190e-01 4.355362e+00
#>  [366] 2.151915e-02 1.162972e+00 4.162542e-01 3.817415e-01 1.149470e+00
#>  [371] 5.775652e-01 6.340898e-02 1.445362e-02 1.266478e-01 6.268852e+00
#>  [376] 3.088927e-01 1.262115e+01 2.690243e+00 1.741696e-01 6.005639e-03
#>  [381] 3.488604e-02 8.274199e-02 3.098347e-01 4.258981e-01 1.549338e-02
#>  [386] 3.703634e+00 5.611853e-01 6.916128e-02 1.549280e-02 2.663043e-02
#>  [391] 3.240819e-01 5.163743e-01 5.675431e-01 3.147109e-02 1.024498e+01
#>  [396] 3.624415e-02 7.654518e-01 9.785491e+00 1.152055e-01 2.668279e-02
#>  [401] 5.005766e-01 2.281945e-01 2.236789e+00 1.054240e-03 8.282880e-01
#>  [406] 8.086732e-01 1.481442e+00 7.461188e+00 1.277127e+00 4.445065e-03
#>  [411] 1.739842e+00 9.192585e+00 5.059593e-01 7.760295e-02 2.687496e+00
#>  [416] 3.510416e-01 1.574038e-01 9.604207e-03 2.003344e+00 2.069903e+00
#>  [421] 4.383027e-03 1.769788e-02 5.223353e-01 1.887688e+00 2.659501e-01
#>  [426] 2.360035e-02 7.494327e-02 6.881146e-01 8.787462e-01 2.174276e+01
#>  [431] 4.711314e+01 4.659459e-01 1.077788e+00 3.962822e-02 9.300163e-01
#>  [436] 4.197356e+00 3.163966e-01 1.889904e+00 1.161579e+00 3.953627e+00
#>  [441] 1.203038e-01 6.394390e-01 1.015518e-01 3.076869e-01 1.459555e+01
#>  [446] 4.726054e-02 1.280060e-01 1.254492e+01 3.298669e-01 8.934089e-02
#>  [451] 5.712107e+00 4.615174e+01 5.561711e+00 2.190381e+00 2.029349e+00
#>  [456] 6.897941e+00 2.814650e-03 6.333572e-02 9.849605e-03 2.704463e-01
#>  [461] 7.367042e-02 3.258439e+00 3.564685e+00 7.102469e+00 9.045191e-01
#>  [466] 5.235148e-01 3.201805e-01 9.413621e-02 3.413631e+01 1.322702e+01
#>  [471] 2.262688e+00 9.641867e+00 6.048530e-02 6.730623e+01 4.934081e-01
#>  [476] 3.620740e+00 4.993428e-01 6.293755e-01 1.111161e-01 1.296759e+00
#>  [481] 2.830361e+01 1.019082e+00 1.193421e+00 2.263032e+00 8.929873e-01
#>  [486] 3.119469e-02 1.976959e-02 3.552434e+00 2.803270e-02 9.472038e-01
#>  [491] 1.625576e+00 4.400506e+00 8.224630e-02 3.278136e-02 3.655616e+00
#>  [496] 4.775572e-01 1.115066e+01 1.366092e-01 1.279240e+00 1.185678e+00
#>  [501] 1.849205e+00 2.423834e+00 2.173745e-01 6.295610e-01 1.250703e+00
#>  [506] 1.045027e-01 1.069984e-04 1.909643e+01 6.730722e-02 7.179875e-01
#>  [511] 7.052783e+00 8.486422e-01 6.067286e+00 9.816695e-01 3.128149e-01
#>  [516] 3.250417e-03 1.942058e-02 1.158712e+02 1.096241e-02 3.226845e+00
#>  [521] 9.770823e-01 1.953953e-01 1.147976e+00 5.798470e+00 3.348259e-01
#>  [526] 7.873860e+00 7.924632e-01 1.552377e+00 1.526062e-01 1.868070e+00
#>  [531] 7.736066e-02 4.888785e-01 1.215901e+00 1.933847e-03 7.169003e-01
#>  [536] 1.040373e+00 9.357715e-01 1.055356e+00 3.179236e+00 3.617845e+00
#>  [541] 5.901798e-01 2.402992e+03 1.118558e+01 4.505521e-01 8.540171e-01
#>  [546] 1.661759e+01 4.088302e-01 1.200679e-01 9.958530e-01 1.872200e+00
#>  [551] 5.988529e-01 3.612150e-01 5.476166e-01 7.396630e-02 1.336846e-01
#>  [556] 7.946746e-01 3.560131e-01 3.788636e+00 1.002587e-02 1.694833e+00
#>  [561] 4.389657e-01 1.023786e-01 8.599724e-01 4.918080e-01 4.325590e+01
#>  [566] 5.623058e-01 9.205493e-01 1.526398e-01 2.552144e+00 4.611942e+01
#>  [571] 2.457957e-01 5.353182e-01 5.576898e-01 6.781247e+00 5.629943e-01
#>  [576] 9.251601e-01 1.802044e-01 1.335510e-02 7.811890e+00 1.316407e-01
#>  [581] 5.120675e-01 2.869761e+00 1.214581e-02 3.080298e-01 5.195046e-01
#>  [586] 8.687921e-01 1.064741e-02 8.891887e-01 4.098805e-02 7.478332e-01
#>  [591] 4.006914e-01 1.156124e+00 6.728043e-02 2.920449e-03 6.281752e-01
#>  [596] 9.049694e-02 3.446879e-03 4.588737e-05 9.693832e-01 1.629994e-01
#>  [601] 3.983041e+00 1.500577e+00 1.393048e+00 9.875514e-02 1.580576e+02
#>  [606] 1.909904e+00 5.179286e-01 2.019349e-01 5.268688e-01 1.981881e+00
#>  [611] 3.893913e+00 2.469547e-01 2.058841e+00 2.463165e-04 1.436666e+00
#>  [616] 1.762551e+00 8.108807e-01 1.470765e-01 9.789014e-01 6.365339e+00
#>  [621] 2.606369e+00 2.223707e+00 1.524441e+00 8.315810e-01 1.344918e-01
#>  [626] 2.588728e+00 5.167730e+00 8.263910e-01 1.652240e+00 5.390664e-03
#>  [631] 1.760385e+02 9.718761e-01 2.111480e+00 1.369913e-01 1.624793e-02
#>  [636] 1.772690e+00 3.570066e-02 2.069716e+00 2.535722e-01 1.583980e+00
#>  [641] 2.554908e+00 7.208932e-01 2.753339e+00 1.384017e+00 1.433798e-01
#>  [646] 1.303285e+00 2.347592e+00 2.719974e+00 1.285601e+00 7.903350e-02
#>  [651] 1.375060e-01 2.509178e-01 3.436971e+00 1.528189e-02 4.897620e-02
#>  [656] 3.793236e-02 1.012459e-01 3.696941e-01 1.742247e-02 3.843813e-03
#>  [661] 4.239626e+00 2.202224e+01 4.886605e+00 2.002471e+00 9.185637e+00
#>  [666] 1.987500e-01 5.096033e-01 7.298243e-01 2.573954e-03 9.727619e-01
#>  [671] 4.609978e-01 2.253804e+00 6.286156e-02 1.826369e-01 5.547731e-01
#>  [676] 6.035843e+00 1.965520e-01 1.032178e-01 1.466847e+00 1.456420e+00
#>  [681] 9.367643e-01 2.592012e+00 5.014948e+00 1.997142e+00 7.406173e-02
#>  [686] 2.974937e-01 5.284082e-01 7.361738e-01 6.467351e-02 6.890162e-01
#>  [691] 1.450098e-01 1.204238e+01 4.005901e-02 8.011994e-02 6.926602e-01
#>  [696] 3.772796e-01 1.352966e+01 8.159815e-02 6.329349e+01 3.485181e-03
#>  [701] 4.985489e-02 2.805842e-01 1.588586e-01 8.254492e-02 3.637279e+02
#>  [706] 1.041622e+01 6.249631e-02 2.600534e+00 5.947081e-02 8.019746e-01
#>  [711] 8.754958e-01 1.614511e-01 4.474081e-02 2.772843e+00 9.033583e-01
#>  [716] 1.261884e+00 3.373497e-01 1.868581e+00 1.496840e+00 1.304480e+00
#>  [721] 2.283433e+00 5.485047e+00 2.307770e+00 1.053686e-02 1.825291e+00
#>  [726] 4.085961e+00 1.294867e+00 1.294888e+00 1.102351e+00 1.290879e+01
#>  [731] 2.389520e-01 1.076084e-01 1.190341e-01 1.678821e-03 8.339413e-02
#>  [736] 1.257232e-01 9.335235e-02 4.210563e+00 1.716427e+00 6.420330e-01
#>  [741] 4.066185e-01 2.282603e-03 6.519673e-01 1.127911e+02 1.273712e-01
#>  [746] 1.121042e+00 5.949643e-01 1.751647e+00 1.464755e-01 2.948046e-01
#>  [751] 1.421125e+00 2.806761e+00 2.606765e+00 3.735190e-03 9.965882e+00
#>  [756] 2.656838e-01 7.860897e-01 1.294922e-01 2.509038e-02 4.300670e+01
#>  [761] 7.446133e-01 4.858952e-01 8.626354e-01 1.971880e-01 4.098605e-01
#>  [766] 1.647184e+00 2.425197e-02 2.125453e+00 3.449718e-01 4.022761e-02
#>  [771] 4.230619e-01 8.443482e-02 1.962122e+00 2.810673e+00 9.044491e+01
#>  [776] 3.539565e-02 7.266430e-02 8.836266e-01 3.186617e+00 2.263603e-01
#>  [781] 4.056632e-01 7.952235e-01 9.024339e-01 2.463373e-01 1.960763e-01
#>  [786] 1.823458e+00 4.022381e-02 1.658353e-02 2.536435e+00 1.442793e-01
#>  [791] 1.505287e+00 9.102968e-02 3.412687e-01 3.406803e-01 3.129662e+00
#>  [796] 1.055432e+00 4.939453e-03 2.700235e-01 2.746193e-02 1.943541e+00
#>  [801] 5.184242e-01 1.211508e-01 9.733760e-01 6.308679e-03 7.268066e-01
#>  [806] 1.854310e+00 1.486455e-01 2.703047e-01 3.118585e-01 6.760749e-01
#>  [811] 9.470546e-01 5.512184e-01 8.867613e-02 8.068200e-01 1.929396e+01
#>  [816] 1.960323e-01 7.480431e-02 5.637931e-01 2.888748e+01 3.945925e-01
#>  [821] 2.109210e+00 1.625379e-01 1.920184e+00 6.241183e+00 9.805397e-02
#>  [826] 4.139123e+00 6.343938e+00 2.866808e+00 2.803962e-01 3.750635e-01
#>  [831] 3.677689e+00 1.928288e+01 2.836886e-01 2.904787e-03 1.574874e+01
#>  [836] 6.822793e-02 1.019113e+00 1.109249e+01 6.274739e+01 2.080661e-01
#>  [841] 3.786183e-01 1.525741e-02 3.824682e-01 4.347208e-01 6.318362e-02
#>  [846] 1.241608e-01 8.580651e-03 3.347876e-01 6.838418e+00 1.371592e+00
#>  [851] 8.100303e-01 6.980369e-02 3.141939e-01 2.168393e+01 2.652042e+00
#>  [856] 1.711379e+00 8.763592e-01 5.839519e-01 2.773766e+00 1.171992e+00
#>  [861] 1.893514e-01 5.052341e-02 2.431827e+00 6.765430e-01 1.536283e+03
#>  [866] 2.403909e-02 7.332175e-01 1.098769e+00 2.238473e+01 1.922998e-02
#>  [871] 6.357491e-01 8.148560e-01 2.021885e-01 1.397954e-02 3.265467e-01
#>  [876] 8.798048e+00 3.117671e-01 7.539305e-01 1.529960e+00 2.208926e+00
#>  [881] 2.242759e-01 2.854570e-02 2.740168e+00 1.150433e+00 7.312965e-01
#>  [886] 2.112164e+00 7.106665e-01 1.180170e+00 3.277342e-01 7.953601e-01
#>  [891] 6.526737e-02 3.031507e-01 1.729924e+01 1.480209e+00 1.791108e+00
#>  [896] 5.135056e-02 1.185686e+00 3.750107e-01 1.307797e-01 8.239605e-01
#>  [901] 1.057366e+00 2.622548e-01 2.283888e+00 1.201196e-01 6.110355e-01
#>  [906] 1.145218e+00 3.416493e-01 5.571317e+00 6.920107e+00 5.026295e-02
#>  [911] 1.408748e+02 2.088516e-02 1.792195e-03 6.909424e-01 4.513306e-01
#>  [916] 3.401309e-03 9.605034e-01 8.337276e-01 2.384282e-01 1.466896e+00
#>  [921] 6.014602e-01 5.243842e+00 2.396799e+00 1.150782e+01 2.349998e-02
#>  [926] 2.000090e-03 1.770963e-01 1.985813e-01 5.351217e-02 1.513306e+00
#>  [931] 4.479900e-01 4.811773e-01 1.859092e-01 6.190843e-01 2.367020e+00
#>  [936] 9.091015e-03 6.230682e-01 3.819889e-01 2.327726e-01 2.433039e-01
#>  [941] 7.155964e-01 2.102849e+00 3.126314e-04 4.815721e-01 8.198513e-02
#>  [946] 4.539992e+00 8.920291e-02 7.986307e-04 9.383552e-01 7.225710e-02
#>  [951] 1.327264e+00 9.627949e-02 2.964908e+00 5.599177e+00 8.288855e-01
#>  [956] 2.671923e+02 2.194643e-01 1.213313e-01 7.323917e-01 2.738630e+00
#>  [961] 5.590235e-03 1.601910e+00 1.166858e+00 1.673258e+00 3.563736e-01
#>  [966] 1.036094e+00 2.138440e+00 3.118397e+00 1.992743e-01 1.244717e+00
#>  [971] 1.037861e+01 8.464225e-01 1.059901e+00 9.502859e-01 1.475757e+00
#>  [976] 1.533571e-02 8.060881e+00 3.767492e+00 1.375231e+00 2.058089e+00
#>  [981] 3.933970e+00 1.318760e+00 4.413829e-01 8.325355e-02 1.375114e-03
#>  [986] 1.360358e+00 5.789420e+00 2.325324e+00 1.378198e+00 8.866063e-01
#>  [991] 1.289032e+00 4.365029e-01 4.958537e-01 4.564915e-01 1.735381e-01
#>  [996] 1.093995e-01 2.018960e-01 3.454284e+00 1.266783e+00 3.146022e+00

## Fitting Mittag-Leffler distribution to observations X by Maximum
## Likelihood

mlml <- function(X) {
  log_l <- function(theta) {
    #transform parameters so can do optimization unconstrained
    theta[1] <- 1/(1+exp(-theta[1]))
    theta[2] <- exp(theta[2])
    -sum(log(dml(X,theta[1],theta[2])))
  }
  ml_theta <- stats::optim(c(0.5,0.5), fn=log_l)$par
  ml_a <- 1/(1 + exp(-ml_theta[1]))
  ml_d <- exp(ml_theta[2])
  print(paste("tail =", ml_a, "scale =", ml_d))
}
mlml(rml(n = 100, tail = 0.9, scale = 2))

#> [1] "tail = 0.934382398038011 scale = 1.84733540481174"